pyproject.toml
[project]
name = "books"
version = "0.1.0"
description = "Kindle等のスクリーンショット連続撮影 & MEGAアップロードツール (Hyprland専用)"
requires-python = ">=3.10,<4"

dependencies = [
    "Pillow>=10,<12",
    "numpy>=1.26,<3",
]

[project.scripts]
books-screenshot = "scripts.screenshot:main"
books-upload = "scripts.upload:main"
books-setup = "scripts.setup:main"

[dependency-groups]
dev = [
    "pytest>=8,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
package = true
environments = ["sys_platform == 'linux'"]
